requirements.txt
PyJWT
websockets
fastapi
uvicorn
redis
aioredis
pydantic